Compress
Removed an unneeded part of an if condition. The condition checks if the top of the snowflake is below the bottom of the screen. The previous condition was checking if the top of the snowflake was below the bottom of the screen + 2*heightOfSnowflake. Once the top of the snowflake is below the bottom of the screen then it can be reset. The extra "padding" is not needed.